+++ pokemon-wiki-web/src/main/java/cn/fantasticmao/pokemon/web/PokemonConfiguration.java
@@ -1,23 +1,12 @@
 package cn.fantasticmao.pokemon.web;
 
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
-import org.springframework.beans.factory.FactoryBean;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.Database;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.sqlite.SQLiteConfig;
 
-import javax.annotation.Resource;
-import javax.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
-import java.util.Map;
 
 /**
  * PokemonConfiguration
@@ -29,38 +18,12 @@ import java.util.Map;
 public class PokemonConfiguration implements WebMvcConfigurer {
     @Value("${app.dbfile:pokemon_wiki.db}")
     private String databaseFile;
-    @Resource
-    private Map<String, String> hibernateProperties;
 
     @Bean
     public DataSource dataSource() {
-        HikariConfig hikariConfig = new HikariConfig();
-        hikariConfig.setJdbcUrl("jdbc:sqlite:" + databaseFile);
-
-        SQLiteConfig sqLiteConfig = new SQLiteConfig();
-        sqLiteConfig.setDateStringFormat("yyyy-MM-dd HH:mm:ss");
-        hikariConfig.setDataSourceProperties(sqLiteConfig.toProperties());
-        return new HikariDataSource(hikariConfig);
-    }
-
-    @Bean
-    public FactoryBean<EntityManagerFactory> entityManagerFactory(DataSource dataSource) {
-        HibernateJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
-        vendorAdapter.setShowSql(false);
-        vendorAdapter.setGenerateDdl(false);
-        vendorAdapter.setDatabase(Database.MYSQL);
-
-        LocalContainerEntityManagerFactoryBean factoryBean = new LocalContainerEntityManagerFactoryBean();
-        factoryBean.setJpaVendorAdapter(vendorAdapter);
-        factoryBean.setPackagesToScan("cn.fantasticmao.pokemon.web.domain");
-        factoryBean.setDataSource(dataSource);
-        factoryBean.setJpaPropertyMap(hibernateProperties);
-        return factoryBean;
-    }
-
-    @Bean
-    public PlatformTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
-        return new JpaTransactionManager(entityManagerFactory);
+        return DataSourceBuilder.create()
+            .url("jdbc:sqlite:" + databaseFile)
+            .build();
     }
 
 }