+++ pokemon-wiki-web/src/main/java/cn/fantasticmao/pokemon/web/repoistory/NatureRepository.java
@@ -1,7 +1,7 @@
 package cn.fantasticmao.pokemon.web.repoistory;
 
-import cn.fantasticmao.mundo.data.jdbc.NativeQuery;
 import cn.fantasticmao.pokemon.web.domain.Nature;
+import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 import javax.annotation.Nonnull;
@@ -16,6 +16,6 @@ import java.util.List;
 public interface NatureRepository extends PagingAndSortingRepository<Nature, Integer> {
 
     @Nonnull
-    @NativeQuery("SELECT * FROM pw_nature")
+    @Query("SELECT * FROM t_nature")
     List<Nature> findAll();
 }