+++ pokemon-wiki-web/src/main/java/cn/fantasticmao/pokemon/web/repoistory/PokemonDetailLearnSetByLevelingUpRepository.java
@@ -1,8 +1,9 @@
 package cn.fantasticmao.pokemon.web.repoistory;
 
-import cn.fantasticmao.mundo.data.jdbc.NativeQuery;
 import cn.fantasticmao.pokemon.web.domain.PokemonDetailLearnSetByLevelingUp;
+import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.data.repository.query.Param;
 
 import java.util.Collection;
 import java.util.List;
@@ -15,6 +16,6 @@ import java.util.List;
  */
 public interface PokemonDetailLearnSetByLevelingUpRepository extends PagingAndSortingRepository<PokemonDetailLearnSetByLevelingUp, Integer> {
 
-    @NativeQuery("SELECT * FROM pw_pokemon_detail_learn_set_by_leveling_up WHERE idx IN ?1")
-    List<PokemonDetailLearnSetByLevelingUp> findByIndexIn(Collection<Integer> ids);
+    @Query("SELECT * FROM t_pokemon_detail_learn_set_by_leveling_up WHERE idx IN (:ids)")
+    List<PokemonDetailLearnSetByLevelingUp> findByIndexIn(@Param("ids") Collection<Integer> ids);
 }